requirements.txt
ansible==8.2.0
jmespath==1.0.1
pytest==7.4.0
PyYAML==6.0.1
